use only 32-bits for encoder position, instead of 64

+++ RotaryEncoder101.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <Arduino.h>
 
-typedef uint64_t encoderPos_t;
+typedef uint32_t encoderPos_t;
 
 class RotaryEncoderInterruptA;
 class RotaryEncoderInterruptB;
